Moved image reduction code from NewsBase to own file
[This is only an intermediary commit that facilitates diffing the changes in the next commit.]
