+++ news/models.py
@@ -1,9 +1,5 @@
-import sys
 import uuid
-from io import BytesIO
 
-from PIL import Image
-from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.db import models
 from django.db.models import Q
 from django.utils import timezone
@@ -11,7 +7,6 @@ from django.utils.translation import gettext_lazy as _
 
 from users.models import User
 from util.locale_utils import short_date_format
-from util.logging_utils import get_request_logger
 from util.storage import OverwriteStorage, UploadToUtils
 from web.modelfields import URLTextField, UnlimitedCharField
 from web.multilingual.modelfields import MultiLingualRichTextUploadingField, MultiLingualTextField
@@ -59,30 +54,6 @@ class NewsBase(models.Model):
 
     def __str__(self):
         return str(self.title)
-
-    def save(self, **kwargs):
-        """
-        Override of save, to change all JPEG images to have quality 90. This greatly reduces the size of JPEG images,
-        while resulting in non to very minuscule reduction in quality. In almost all cases, the possible reduction in
-        quality will not be visible to the naked eye.
-        """
-        # Only check the image if there is actually an image
-        if self.image:
-            # PIL will throw an IO error if it cannot open the image, or does not support the given format
-            try:
-                image = Image.open(self.image)
-                if image.format == "JPEG":
-                    output = BytesIO()
-                    image.save(output, format="JPEG", quality=90)
-                    output.seek(0)
-
-                    self.image = InMemoryUploadedFile(output, "ImageField", self.image.name, "image/jpeg",
-                                                      sys.getsizeof(output), None)
-                # Should not close image, as Django uses the image and closes it by default
-            except IOError as e:
-                get_request_logger().exception(e)
-
-        super().save(**kwargs)
 
 
 class ArticleQuerySet(NewsBaseQuerySet):
